+++ test/bigtest/tests/resource-show-test.js
@@ -5,6 +5,7 @@ import setupApplication, { axe } from '../helpers/setup-application';
 import ResourcePage from '../interactors/resource-show';
 import PackageEditPage from '../interactors/package-edit';
 import { entityAuthorityTypes } from '../../../src/constants';
+// import usageConsolidationInfoPopoverTests from './usage-consolidation-info-popover';
 
 describe('ResourceShow', () => {
   setupApplication();
@@ -304,6 +305,9 @@ describe('ResourceShow', () => {
           expect(ResourcePage.usageConsolidationSection.content.summaryTable.rows(0).cells(2).content).to.equal('$0.04 (USD)');
         });
       });
+
+      // TODO:: uncomment current tests after folio/stripes 5.1.0 will be available
+      // usageConsolidationInfoPopoverTests(ResourcePage.usageConsolidationSection.infoPopover);
     });
 
     describe('when token is not needed', () => {